changed type of withdrawn

+++ internal/models/balance/balance.go
@@ -2,5 +2,5 @@ package balance
 
 type Balance struct {
 	Current   float32 `json:"current"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float32 `json:"withdrawn"`
 }
